Support x64 architecture (both Debug and Release modes)
- Generated glfw, assimp, soil2 and freetype x64 libraries and stored them in a specific x64 folder,
- Dragged the previously used libraries in a x32 folder,
- Modified the project properties (Configuration Properties VC++ Directories - includes and libraries, as well as Linker Inputs - Additional Dependencies) to update the relative SolutionDir paths,
- Modified a bunch of casts (use size_t instead of unsigned int) to get rid of all the x64 warnings,
- Solved the 'assimp library not found' issue when clicking on the executable by copy-pasting the dll in the same folder as the executable.
- Still executable issues to be solved (relative Shader paths not working, and still this white window that disappears after 3 secs...)

+++ OpenGLProject/OpenGLProject/Belt.cpp
@@ -94,7 +94,7 @@ void Belt::Render(const Renderer& renderer, const unsigned int& textureUnit)
 
 	for (unsigned int i = 0; i < instanceModel.GetMeshes().size(); ++i)
 	{
-		renderer.DrawInstanced(vao, instanceModel.GetMeshes()[i].GetIndicesCount(), instanceModelNb);
+		renderer.DrawInstanced(vao, static_cast<unsigned int>(instanceModel.GetMeshes()[i].GetIndicesCount()), instanceModelNb);
 	}
 }
